Login and Registration
This takes care of logging in and registration. Ensures form validation for input that its in correct format. If user exists, logs in and starts session. If not, registration checks if user exists, if not, adds new user and redirects to log in. Files home.php, registration.php ensure that there are no running sessions, if so, they're loggedout and redirected to log in. login.php also ensures all sessions are destroyed prior to starting new.